backend/requirements.txt
# Backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
pdf2image==1.17.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
easyocr==1.7.1
ultralytics==8.1.0
torch==2.2.0
torchvision==0.17.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
